# version/server/Server/api/openapi.yaml
openapi: 3.0.1
info:
  title: Swagger ToDo Manager
  description: REST APIs for the interaction with the ToDo Manager service.
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
- url: https://todo.swagger.io/v2
- url: http://todo.swagger.io/v2
...
paths:
  /api/tasks/public:
    get:
      tags:
      - tasks/public
      summary: Retrieve the public tasks
      description: "The public tasks (i.e., the tasks that are visible for all the\
        \ users of the service) are retrieved. This operation does not require authentication.\
        \ A pagination mechanism is implemented to limit the size of messages."
      operationId: getPublicTasks
      parameters:
      - name: pageNo
        in: query
        description: "The id of the requested page (if absent, the first page is returned)"
        required: false
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      x-swagger-router-controller: Taskspublic
  /api/tasks:
    post:
      tags:
      - tasks
      summary: Create a new task
      description: A new task is create by an authenticated user
      operationId: addTask
      requestBody:
        description: Parameters needed to create a new task
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
                x-content-type: application/json
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Tasks
  /api/tasks/{taskId}:
    get:
      tags:
      - "tasks/{taskId}"
      summary: Get a task
      description: "The user can be retrieve an his task by taskId if at least one\
        \ of the following conditions is satisfied: the task is public, the user is\
        \ the owner of the task, the user is an assignee of the task"
      operationId: getTask
      parameters:
      - name: taskId
        in: path
        description: ID of the task to retrieve
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
                x-content-type: application/json
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: TaskstaskId
    put:
      tags:
      - "tasks/{taskId}"
      summary: Update a task
      description: The task with ID taskId is updated. This operation does not allow
        to mark a task as completed. This operation can be performed only by the owner.
      operationId: updateTask
      parameters:
      - name: taskId
        in: path
        description: ID of the task to update
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: The updated task object that needs to replace the old object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        required: true
      responses:
        "204":
          description: No Content
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: TaskstaskId
    delete:
      tags:
      - "tasks/{taskId}"
      summary: Delete a task
      description: The task with ID taskId is deleted. This operation can only be
        performed by the owner.
      operationId: deleteTask
      parameters:
      - name: taskId
        in: path
        description: ID of the task to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: TaskstaskId
  /api/users:
    get:
      tags:
      - users
      summary: Get information about the users
      description: The available information (passwords excluded) about all the users
        is retrieved.
      operationId: getUsers
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
                x-content-type: application/json
        "401":
          description: Unauthorized
          content: {}
        "404":
          description: Not Found
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Users
  /api/users/{userId}:
    get:
      tags:
      - "users/{userId}"
      summary: Get information about a user
      description: The available information (password excluded) about the user specified
        by userId is retrieved.
      operationId: getSingleUser
      parameters:
      - name: userId
        in: path
        description: ID of the user to get
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401":
          description: Unauthorized
          content: {}
        "404":
          description: Not Found
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: UsersuserId
  /api/users/authenticator:
    post:
      tags:
      - users/authenticator
      summary: Logs a user in or out
      description: The user who wants to log in or out sends the user data to the
        authenticator which performs the operation.
      operationId: authenticateUser
      parameters:
      - name: type
        in: query
        description: The operation type ("login" or "logout")
        required: false
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        description: "The data of the user who wants to perform log in or log out.\
          \ For login the structure must contain email and password. For logout, the\
          \ structure must contain the user id."
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "204":
          description: No Content
          headers:
            Set-Cookie:
              style: simple
              explode: false
              schema:
                type: string
                example: token=abcde12345; Path=/; HttpOnly
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized - Wrong password
          content: {}
        "404":
          description: Not found - Wrong username
          content: {}
      x-swagger-router-controller: Usersauthenticator
...
  /api/tasks/{taskId}/completion:
    put:
      tags:
      - "tasks/{taskId}/completion"
      summary: Complete a task
      description: The task with ID taskId is marked as completed. This operation
        can only be performed by an assignee of the task.
      operationId: completeTask
      parameters:
      - name: taskId
        in: path
        description: ID of the task
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: TaskstaskIdcompletion
  /api/users/{userId}/tasks/created:
    get:
      tags:
      - "users/{userId}/tasks/created"
      summary: Retrieve the tasks created by the user with ID userId
      description: This operation allows an authenticated user to retrieve the tasks
        that she has created. A pagination mechanism is implemented to limit the size
        of messages.
      operationId: getUserTasks
      parameters:
      - name: userId
        in: path
        description: ID of the user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      - name: pageNo
        in: query
        description: "The id of the requested page (if absent, the first page is returned)"
        required: false
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: UsersuserIdtaskscreated
  /api/users/{userId}/tasks/assigned:
    get:
      tags:
      - "users/{userId}/tasks/assigned"
      summary: Retrieve the tasks assigned to the user with ID userId
      description: This operation allows an authenticated user to retrieve the tasks
        that has been assigned to her. A pagination mechanism is implemented to limit
        the size of messages.
      operationId: getAssignedTasks
      parameters:
      - name: userId
        in: path
        description: ID of the user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      - name: pageNo
        in: query
        description: "The id of the requested page (if absent, the first page is returned)"
        required: false
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: UsersuserIdtasksassigned
...
components:
  schemas:
    Task:
      type: object
      properties:
        $schema:
          type: string
          description: path to the JSON Schema to use for the validation of the JSON
            file describing the task
        id:
          type: integer
          description: unique identifier for the task
        description:
          maxLength: 160
          type: string
          description: textual task description
        important:
          type: boolean
          description: "this property is set to true if the task is marked as important,\
            \ false otherwise"
          default: false
        private:
          type: boolean
          description: "this property is set to true if the task is marked as private,\
            \ false otherwise"
          default: true
        project:
          type: string
          description: project in which the task is inserted
        deadline:
          type: string
          description: the due date and hour of the task
          format: date-time
        completed:
          type: boolean
          description: "this property is set to true if the task is marked as completed,\
            \ false otherwise"
          default: false
        owner:
          $ref: '#/components/schemas/User'
        assignees:
          type: array
          items:
            $ref: '#/components/schemas/User'
        self:
          type: string
          description: self link
          format: uri-reference
      example:
        important: false
        owner:
          password: password
          $schema: $schema
          name: name
          self: self
          id: 5
          email: ""
        private: true
        $schema: $schema
        description: description
        project: project
        assignees:
        - null
        - null
        self: self
        id: 5
        completed: false
        deadline: 2000-01-23T04:56:07.000+00:00
    User:
      type: object
      properties:
        $schema:
          type: string
          description: path to the JSON Schema to use for the validation of the JSON
            file describing the task
        id:
          type: integer
          description: unique identifier for the user
        name:
          type: string
          description: username of the user
        email:
          type: string
          description: email address of the user
          format: email
        password:
          maxLength: 20
          minLength: 6
          type: string
          description: password of the user
        self:
          type: string
          description: self link
          format: uri-reference
      example:
        password: password
        $schema: $schema
        name: name
        self: self
        id: 5
        email: ""
    inline_response_200:
      type: object
      properties:
        totalPages:
          type: integer
        currentPage:
          type: integer
        totalItems:
          type: integer
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        next:
          type: string
      example:
        next: next
        totalItems: 1
        totalPages: 0
        currentPage: 6
        tasks:
        - important: false
          owner:
            password: password
            $schema: $schema
            name: name
            self: self
            id: 5
            email: ""
          private: true
          $schema: $schema
          description: description
          project: project
          assignees:
          - null
          - null
          self: self
          id: 5
          completed: false
          deadline: 2000-01-23T04:56:07.000+00:00
        - important: false
          owner:
            password: password
            $schema: $schema
            name: name
            self: self
            id: 5
            email: ""
          private: true
          $schema: $schema
          description: description
          project: project
          assignees:
          - null
          - null
          self: self
          id: 5
          completed: false
          deadline: 2000-01-23T04:56:07.000+00:00
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: JWT to identify the authenticated user.
      name: connect.sid
      in: cookie